+++ src/common/sdf.glsl
@@ -1,0 +1,8 @@
+float sdfSphere(vec3 uv, vec4 sphere) {
+  return length(uv - sphere.xyz) - sphere.w;
+}
+
+float sdfBox(vec3 p, vec3 b) {
+  vec3 q = abs(p) - b;
+  return length(max(q, 0.0)) + min(max(q.x, max(q.y, q.z)), 0.0);    
+}